backend/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
grad-cam
numpy
pydantic-settings
python-dotenv
python-multipart
albumentations
scikit-learn
